DX-914: Refactor Jekyll page processing
Refactor Jekyll page processing into a new JekyllPageProcessor class.
